refer functionity

+++ src/main/java/com/example/demo/repository/UserDao.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entity.Users;
 
 @Repository
-public interface UserDao extends JpaRepository<Users, Integer> {
+public interface UserDao extends JpaRepository<Users, Long> {
 
 	public List<Users> findByMobile(Long mobile);
 }
